Update file configuration and add some unit tests.
