+++ .travis.yml
@@ -74,7 +74,7 @@ script:
 after_success:
   - if [ "$CXX" = "g++"                 ]; then lcov --directory . --capture --output-file coverage.info; fi
   - if [ "$CXX" = "clang++-${LLVM_VER}" ]; then lcov --directory . --capture --output-file coverage.info --gcov-tool llvm-cov-${LLVM_VER}; fi
-  - lcov --remove coverage.info 'src/renderer/*' 'src/viewer/*' 'test/*' 'example/*' '/usr/*' 'CMakeFiles/*' --output-file coverage.info
+  - lcov --remove coverage.info '3rdparty/*' 'src/renderer/*' 'src/viewer/*' 'test/*' 'example/*' '/usr/*' 'CMakeFiles/*' --output-file coverage.info
   - lcov --list coverage.info
   - coveralls-lcov --repo-token RiYcPJSCbPZoogMd1PE10696EAqG8sl5q coverage.info
