+++ include/spica.h
@@ -10,7 +10,6 @@
 #include "../src/utils/timer.h"
 #include "../src/utils/vector3d.h"
 #include "../src/utils/sampler.h"
-#include "../src/utils/vbo.h"
 #include "../src/utils/spica_dirs.h"
 
 #include "../src/geometry/geometry.h"
