+++ src/utils/CMakeLists.txt
@@ -8,7 +8,6 @@ configure_file(${CMAKE_CURRENT_LIST_DIR}/spica_dirs.h.in
      ${CMAKE_CURRENT_LIST_DIR}/color.cc
      ${CMAKE_CURRENT_LIST_DIR}/image.cc
      ${CMAKE_CURRENT_LIST_DIR}/sampler.cc
-     ${CMAKE_CURRENT_LIST_DIR}/vbo.cc
      PARENT_SCOPE)
 
  set(HEADERS ${HEADERS}
@@ -20,7 +19,6 @@ configure_file(${CMAKE_CURRENT_LIST_DIR}/spica_dirs.h.in
      ${CMAKE_CURRENT_LIST_DIR}/color.h
      ${CMAKE_CURRENT_LIST_DIR}/image.h
      ${CMAKE_CURRENT_LIST_DIR}/sampler.h
-     ${CMAKE_CURRENT_LIST_DIR}/vbo.h
      ${CMAKE_CURRENT_LIST_DIR}/timer.h
      ${CMAKE_CURRENT_LIST_DIR}/axis_comparable.h
      ${CMAKE_CURRENT_LIST_DIR}/kdtree.h
